\name{genepop-internals}
\alias{genepop-internals}

\alias{summary.GP_outfile}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal genepop Functions}
\description{
  These are not to be called by the user, or are waiting for documentation to be written.
}

\keyword{ internal }