% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genepop.R, R/RcppExports.R
\name{set_restriction}
\alias{set_restriction}
\alias{genepop-utils}
\alias{getVersion}
\title{Programming utilities}
\usage{
set_restriction(set = FALSE)

getVersion()
}
\arguments{
\item{set}{logical: whether to set restrictions on number of populations and of loci}
}
\description{
\code{getVersion} returns the version number of the C++ code (the same number that identifies the C++ executable). \code{set_restriction(TRUE)} sets the maximum number of populations and of loci to 300.
}
