% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genepop.R
\name{genepopExample}
\alias{genepopExample}
\title{Copy an example file from the Genepop package distribution}
\usage{
genepopExample(filename)
}
\arguments{
\item{filename}{The name of an example file from the Genepop distribution.}
}
\value{
Returns the \code{filename}
}
\description{
This function is used to copy an example file to the user's directory. It should not be used when analysing one's own data!
}
