% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setMantelSeed}
\alias{setMantelSeed}
\title{Set random generator seed for Mantel test}
\usage{
setMantelSeed(seed)
}
\arguments{
\item{seed}{integer: the new seed}
}
\description{
Set random generator seed for Mantel test
}
