% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setRandomSeed}
\alias{setRandomSeed}
\title{Set random generator seed (except for Mantel test)}
\usage{
setRandomSeed(seed)
}
\arguments{
\item{seed}{integer: the new seed}
}
\description{
Set random generator seed (except for Mantel test)
}
